+++ docs/changelog.rst
@@ -8,7 +8,8 @@ Changelog
 19.11.2
 -------
 
-* new: statistics tracking on WAMP serializers
+* new: statistics tracking on WAMP serializers :class:`autobahn.wamp.serializer.Serializer`
+* new: helper :func:`autobahn.util.time_ns`
 
 19.11.1
 -------